Tighten up conditionals in bulk method
No need for 2-conditional depth.

Rubocop-influenced changes for Hash literals, Style/Next, etc.

+++ app/controllers/catalog_controller.rb
@@ -285,34 +285,25 @@ class CatalogController < ApplicationController
   # Given a directory with bulk metadata upload information (written by ModsulatorJob), loads the job data into a hash.
   def bulk_job_metadata(dir)
     success = 0
-    job_info = Hash.new
+    job_info = {}
     log_filename = File.join(dir, Argo::Config.bulk_metadata_log)
-    if (File.directory?(dir) && File.readable?(dir))
-      if (File.exist?(log_filename) && File.readable?(log_filename))
-        File.open(log_filename, 'r') { |log_file|
-          log_file.each_line do |line|
+    if File.directory?(dir) && File.readable?(dir) && File.exist?(log_filename) && File.readable?(log_filename)
+      File.open(log_filename, 'r') { |log_file|
+        log_file.each_line do |line|
 
-            # The log file is a very simple flat file (whitespace separated) format where the first token denotes the
-            # field/type of information and the rest is the actual value.
-            matched_strings = line.match(/^([^\s]+)\s+(.*)/)
-            if (matched_strings && matched_strings.length == 3)
-              job_info[matched_strings[1]] = matched_strings[2]
-
-              if(matched_strings[1] == 'argo.bulk_metadata.bulk_log_job_save_success')
-                success += 1
-              end
-
-              if(@@error_messages.include?(matched_strings[1]))
-                job_info['error'] = 1
-              end
-            end
-          end
-          job_info['dir'] = get_leafdir(dir)
-          job_info['argo.bulk_metadata.bulk_log_druids_loaded'] = success
-        }
-      end
+          # The log file is a very simple flat file (whitespace separated) format where the first token denotes the
+          # field/type of information and the rest is the actual value.
+          matched_strings = line.match(/^([^\s]+)\s+(.*)/)
+          next unless matched_strings && matched_strings.length == 3
+          job_info[matched_strings[1]] = matched_strings[2]
+          success += 1 if matched_strings[1] == 'argo.bulk_metadata.bulk_log_job_save_success'
+          job_info['error'] = 1 if @@error_messages.include?(matched_strings[1])
+        end
+        job_info['dir'] = get_leafdir(dir)
+        job_info['argo.bulk_metadata.bulk_log_druids_loaded'] = success
+      }
     end
-    return job_info
+    job_info
   end
 
   # Given a DRUID, loads any metadata bulk upload information associated with that DRUID into a hash.
